syscalls: add proper cast + FIXME comment
JIRA: DO-3

+++ syscalls.c
@@ -297,7 +297,7 @@ int syscalls_beginthreadex(u8 *ustack)
 {
 	process_t *proc = proc_current()->process;
 	startFn_t start;
-	unsigned int priority, stacksz;
+	unsigned int priority, stacksz; /* FIXME: stacksz should probably be size_t */
 	void *stack, *arg;
 	int *id;
 	int err;
@@ -319,7 +319,7 @@ int syscalls_beginthreadex(u8 *ustack)
 
 	proc_get(proc);
 
-	err = proc_threadCreate(proc, start, id, (u8)priority, (size_t)SIZE_KSTACK, stack, stacksz, arg);
+	err = proc_threadCreate(proc, start, id, (u8)priority, (size_t)SIZE_KSTACK, stack, (size_t)stacksz, arg);
 
 	if (err < 0) {
 		(void)proc_put(proc);
